need update item

+++ controllers/shopController.js
@@ -11,6 +11,7 @@ const {PutObjectCommand, S3Client} = require('@aws-sdk/client-s3')
 
 
 
+
 const s3Client = new S3Client({
     region: 'us-east-1',
     credentials: {
@@ -48,34 +49,71 @@ exports.create_item_get = asyncHandler(async (req, res, next) => {
     })
 })
 
-exports.create_item_post = asyncHandler(async (req, res, next) => {
-    const bucketName = 'invenappbucket'
-    
-    const imageFile = req.params.file
-    const keyName = `${generateRandomStr(5)}_${generateTimeStamp()}`
-    const params = {
-        Bucket: bucketName,
-        Key: `images/${keyName}.webp`,
-        Body: fs.createReadStream(req.file.path),
-        ACL: 'public-read',
-        
-    }
-    const command = new PutObjectCommand(params)
-    const response = await s3Client.send(command)
-    debug('Image upload successfully')
-    debug('url:', `https://${bucketName}.s3.us-east-1.amazonaws.com/${params.Key}`)
-    const url = `https://${bucketName}.s3.us-east-1.amazonaws.com/${params.Key}`
-    fs.unlink(req.file.path, (err) => {
-        if (err) {
-            console.error('error deleting file')
-        } else {
-            console.error('file deleted succesfully')
-        }
-    })
+exports.create_item_post = [
 
-    res.redirect('/shop')
-      
-})
+    body('itemName', 'Name must not be empty.')
+    .trim()
+    .isLength({min: 1})
+    .escape(),
+    body('itemDescription', 'Must have a description')
+    .trim()
+    .isLength({min: 1})
+    .escape(),
+
+    asyncHandler(async (req, res, next) => {
+
+    const errors = validationResult(req)
+    if (!errors.isEmpty()){
+        const allCategory = await Category.find()
+
+        res.render('item_form', {
+            title:'Create Item',
+            categories: allCategory
+        })
+    } else {
+
+        const bucketName = 'invenappbucket'
+    
+        const imageFile = req.params.file
+        const keyName = `${generateRandomStr(5)}_${generateTimeStamp()}`
+        const params = {
+            Bucket: bucketName,
+            Key: `images/${keyName}`,
+            Body: fs.createReadStream(req.file.path),
+            ACL: 'public-read',
+            
+        }
+        //req.file.path created from multer, fs createreadstream reads it then unlinks removed the file after it's temp stored in the server uploads/ directory set up in routes
+        const command = new PutObjectCommand(params)
+        const response = await s3Client.send(command)
+        debug('Image upload successfully')
+        debug('url:', `https://${bucketName}.s3.us-east-1.amazonaws.com/${params.Key}`)
+        const url = `https://${bucketName}.s3.us-east-1.amazonaws.com/${params.Key}`
+        fs.unlink(req.file.path, (err) => {
+            if (err) {
+                console.error('error deleting file')
+            } else {
+                console.error('file deleted succesfully')
+            }
+        })
+
+        if (!Array.isArray(req.body.category)){
+            req.body.category = [req.body.category]
+        }
+
+        const item = new Item({
+            name: req.body.itemName,
+            description: req.body.itemDescription,
+            price: req.body.itemPrice,
+            number_in_stock: req.body.itemStock,
+            category: req.body.category,
+            imgUrl: url
+        })
+        await item.save()
+        res.redirect('/shop')
+    }
+    
+})]
 
 exports.itemDetails = asyncHandler(async (req, res, next) => {
     const item = await Item.findById(req.params.id).populate('category')
@@ -85,4 +123,11 @@ exports.itemDetails = asyncHandler(async (req, res, next) => {
     } )
 })
 
+exports.item_edit_get = asyncHandler(async (req, res, next) => {
+    const item = await Item.findById(req.params.id).populate('category')
+
+    res.render('item_update_form', {
+        item: item
+    })
+})
